feat: project setup and configure
This project is invoice management system for the particular shop or the seller. Any shopkeeper or salesman can use for their shop and for the purchases product. Here stored data of the products for selling and stored data of buyers or customers. In this application easily identify the stock products and sold prodructs.

#2